updating code for variable/fuction renaming, and incoporating new carbon data

+++ luto/economics/agricultural/revenue.py
@@ -17,21 +17,7 @@
 # You should have received a copy of the GNU General Public License along with
 # LUTO2. If not, see <https://www.gnu.org/licenses/>.
 
-# Copyright 2023 Fjalar J. de Haan and Brett A. Bryan at Deakin University
-#
-# This file is part of LUTO 2.0.
-# 
-# LUTO 2.0 is free software: you can redistribute it and/or modify it under the
-# terms of the GNU General Public License as published by the Free Software
-# Foundation, either version 3 of the License, or (at your option) any later
-# version.
-# 
-# LUTO 2.0 is distributed in the hope that it will be useful, but WITHOUT ANY
-# WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
-# A PARTICULAR PURPOSE. See the GNU General Public License for more details.
-# 
-# You should have received a copy of the GNU General Public License along with
-# LUTO 2.0. If not, see <https://www.gnu.org/licenses/>. 
+ 
 
 """
 Pure functions to calculate economic profit from land use.
@@ -62,7 +48,7 @@ def get_rev_crop( data: Data   # Data object.
     """
     # Check if land-use exists in AGEC_CROPS (e.g., dryland Pears/Rice do not occur), if not return zeros
     if lu not in data.AGEC_CROPS['P1', lm].columns:
-        rev_t = np.zeros((data.NCELLS))
+        rev_t = np.zeros((data.NCELLS)).astype(np.float32)
         
     else:
         rev_multiplier = 1
@@ -118,7 +104,7 @@ def get_rev_lvstk( data: Data   # Data object.
                             )  
 
         # Set Wool and Milk to zero as they are not produced by beef cattle
-        rev_wool = rev_milk = np.zeros((data.NCELLS))
+        rev_wool = rev_milk = np.zeros((data.NCELLS)).astype(np.float32)
 
     elif lvstype == 'SHEEP':    
 
@@ -144,7 +130,7 @@ def get_rev_lvstk( data: Data   # Data object.
                                 )
 
         # Set Milk to zero as it is not produced by sheep
-        rev_milk = np.zeros((data.NCELLS)) # Set Milk to zero as it is not produced by sheep
+        rev_milk = np.zeros((data.NCELLS)).astype(np.float32) # Set Milk to zero as it is not produced by sheep
 
 
     elif lvstype == 'DAIRY':
@@ -158,7 +144,7 @@ def get_rev_lvstk( data: Data   # Data object.
                                 )
 
         # Set Meat, Wool and Live exports to zero
-        rev_meat = rev_wool = rev_lexp = np.zeros((data.NCELLS)) 
+        rev_meat = rev_wool = rev_lexp = np.zeros((data.NCELLS)).astype(np.float32)
 
     else:  # Livestock type is unknown.
         raise KeyError(f"Unknown {lvstype} livestock type. Check `lvstype`.")   
@@ -194,7 +180,7 @@ def get_rev( data: Data    # Data object.
         return get_rev_lvstk(data, lu, lm, yr_idx)
 
     elif lu in data.AGRICULTURAL_LANDUSES:
-        return  pd.DataFrame(np.zeros((data.NCELLS, 1)),
+        return  pd.DataFrame(np.zeros((data.NCELLS, 1)).astype(np.float32),
                              columns=pd.MultiIndex.from_product([[lu],[lm],['Revenue']]))
 
     else:
